power.nb: Power and Sample Size Calculation for Differential Abundance Microbiome Studies

Provides functions for estimating statistical power and required sample sizes in differential abundance microbiome studies using negative binomial models. The methods are based on Agronah and Bolker (2025) doi:10.1371/journal.pone.0318820 . The package includes tools for simulation-based power analysis and sample size estimation using generalized additive models (GAMs), and visualization utilities for exploring the relationship between power, effect size, abundance, and sample size.

See also

Author

Maintainer: Michael Agronah michaelagronah@gmail.com (ORCID)

Authors:

Ben Bolker bolker@mcmaster.ca (ORCID)