Simulate Log Means for OTUs

This function generates log means for a specified number of OTUs (Operational Taxonomic Units) based on the provided parameters. If a single mean is specified, the log means are drawn from a normal distribution. If multiple means and corresponding weights are specified, the log means are drawn from a mixture of normal distributions.

Usage

logmean_sim_fun(logmean_param, notu)

Arguments

logmean_param

A list containing the parameters for the distribution:

mu: A single value or a vector of mean(s) for the normal or mixture distribution.

sigma: The standard deviation(s) for the normal or mixture distribution.

lambda: (Optional) A vector of weights for the components of the mixture distribution. Required only if mu has more than one value.

notu

An integer specifying the number of OTUs to simulate.

Value

A numeric vector of simulated log means for the specified number of OTUs.

Examples

# Example 1: Single normal distribution
params_single <- list(mu = 0, sigma = 1)
logmean_sim_fun(logmean_param = params_single, notu = 100)
#>   [1]  0.175385360  1.111811631  0.702392348 -0.667554028  0.127859830
#>   [6] -0.357975022  0.125561424 -0.453126163  1.069757479 -0.635486708
#>  [11] -0.100971387 -0.065774785 -2.410192597 -2.158413678  0.746719377
#>  [16] -1.256659833  0.567119412  0.173220358 -1.262341614  0.166360618
#>  [21]  0.383822276  0.187167595  1.241310898 -1.148992769 -0.146460224
#>  [26] -0.470902519  1.049591662 -0.925124211 -0.427850623  0.901765985
#>  [31] -2.496688561 -0.949917861  1.249409153  0.196176218  1.235065791
#>  [36] -0.048587118 -0.338929073  0.215309993  0.057692614 -1.185634351
#>  [41]  0.005191341  0.374045951  0.821315754  0.366582819 -1.004199785
#>  [46]  0.912295355 -0.446229447 -0.930216236 -0.316693672  1.177465936
#>  [51] -0.906789432 -1.548335311 -1.734189417  0.527090815 -0.807281074
#>  [56] -0.915889267 -1.201261983 -1.229628574 -0.679163756 -0.049070574
#>  [61] -0.990876598 -0.463628924 -1.336930893  0.378268684 -1.070461028
#>  [66]  1.775370217  0.750846397 -0.835658307  1.910908520 -0.440944844
#>  [71]  1.419119319 -0.891577568  0.388398551  0.937050146 -0.355459769
#>  [76]  0.501265408  0.959158353 -0.067989783  0.003995981 -1.554813884
#>  [81] -0.161846873 -0.463544754 -0.974459529  0.159184104  1.598035291
#>  [86] -0.066878354 -0.334624219 -0.773298218  0.131130235  0.020541102
#>  [91]  0.369400911  0.922275683 -1.329875489  0.931310228 -0.895455449
#>  [96] -1.217318404  0.555168698 -0.888918960  0.694939113 -1.005599499

# Example 2: Mixture of normal distributions
params_mixture <- list(
  mu = c(-1, 1),
  sigma = c(0.5, 0.5),
  lambda = c(0.4, 0.6)
)
logmean_sim_fun(logmean_param = params_mixture, notu = 100)
#>   [1] -0.585249553  0.716450847 -0.566685433 -1.422634118  1.608604797
#>   [6] -0.835797754  2.217829775  0.907618319 -1.406395754  1.331965855
#>  [11]  0.677014902  2.267864700  0.316848706  0.133531541  1.826665215
#>  [16] -0.453888469  0.884762532  0.819617653 -0.069501657 -1.339465260
#>  [21]  1.284260622  1.007216322 -0.689872359 -2.003603236 -1.342147519
#>  [26]  0.375963653 -0.677168849  0.563300401 -1.117747545  1.811570999
#>  [31] -0.529715318 -0.751977088  1.041669578  0.103878484  1.642065271
#>  [36]  1.263632370  2.134514020 -1.655221712  1.249856125 -1.025143823
#>  [41] -0.600560779  1.585416926  1.382024670  1.056041525 -0.434987983
#>  [46]  0.937351132  1.593076136 -0.547154586 -0.174515144  0.496965382
#>  [51] -0.214220488  1.360587659  1.395343720  0.650374252  0.869700737
#>  [56] -0.724349184  1.382380694 -1.424813324 -1.197548633  1.096349929
#>  [61]  1.945570545  1.221663030  1.346224219  1.786412581  1.618009873
#>  [66]  0.312581876  0.855722624  0.321718778 -1.747578074  0.935826899
#>  [71]  1.406824621 -1.128362483 -1.372236805  0.803239386 -0.909748088
#>  [76]  1.445489210 -0.891208428  0.490432552  2.063131982  0.452566407
#>  [81]  0.323561101 -0.928976351 -1.054615734  1.031247377  1.316433500
#>  [86]  0.609242107 -0.024231174  1.001528334  0.407167781  1.902040451
#>  [91]  1.229741632  0.001605662  0.527561052  0.594420335  1.047004643
#>  [96]  0.517421948  1.142257221  0.972923665  0.461508980  0.345987776