Simulate Log Means for OTUs

This function generates log means for a specified number of OTUs (Operational Taxonomic Units) based on the provided parameters. If a single mean is specified, the log means are drawn from a normal distribution. If multiple means and corresponding weights are specified, the log means are drawn from a mixture of normal distributions.

Usage

logmean_sim_fun(logmean_param, notu)

Arguments

logmean_param

A list containing the parameters for the distribution:

mu: A single value or a vector of mean(s) for the normal or mixture distribution.

sigma: The standard deviation(s) for the normal or mixture distribution.

lambda: (Optional) A vector of weights for the components of the mixture distribution. Required only if mu has more than one value.

notu

An integer specifying the number of OTUs to simulate.

Value

A numeric vector of simulated log means for the specified number of OTUs.

Examples

# Example 1: Single normal distribution
params_single <- list(mu = 0, sigma = 1)
logmean_sim_fun(logmean_param = params_single, notu = 100)
#>   [1]  1.5247663842 -0.5015504837  2.3992551103 -0.0959089860  0.0658688799
#>   [6] -2.0950705492  0.0981670854  1.7746868241 -0.0115041725  1.2632880269
#>  [11] -0.9321132804 -0.2760755436 -0.4886184524  0.4711262332  1.4441135506
#>  [16]  0.1991474290  0.8888998756  0.5532362418  1.2451872751  0.2936839672
#>  [21] -0.7746829596  0.0266178681 -2.4446884116 -1.6230134410 -0.5463497414
#>  [26]  0.4292152775  1.3620103995 -0.9161524195  0.0359557023  0.9865549323
#>  [31] -1.0184345010  0.3973988722 -0.7131125048  1.3450399517  0.2042207959
#>  [36]  0.2619146139 -0.4147417530 -1.3658556585  0.8185952324 -0.3744949175
#>  [41]  0.3850703591 -1.3585343837 -0.4399398950 -0.8149376733  0.4604441468
#>  [46]  0.7632444146 -1.1145291638 -1.0696257271  0.1269307895  0.4773923798
#>  [51] -1.0904656577  0.1217674695 -0.5755452000 -0.1922327562  0.1152379157
#>  [56] -0.3157208618 -0.4386662687 -2.3279713000 -0.2114882804  0.2681199832
#>  [61] -0.7841378517 -0.0821148655 -0.1634936795  1.2592703625 -0.6587060791
#>  [66] -1.0064437262 -0.1209984078 -1.3317281620 -0.4359405831 -0.5295206910
#>  [71]  0.1116011738 -0.0868561787 -0.0183377960 -0.4001851323 -0.3694184639
#>  [76] -0.6056917488  1.1312555755  0.0126410777 -0.9858803835 -1.5539200460
#>  [81] -0.4449745407 -0.1227002719 -1.3042977144  0.2100949530  0.4276912705
#>  [86] -1.9329924876  1.7888001957 -0.6517847182 -0.5517179091 -0.9803206663
#>  [91]  0.0003794807 -1.0683507606  0.5574898852 -0.3847882409 -0.0070969234
#>  [96]  0.2714161280 -2.6420007953  0.7687061725 -0.7280291431  0.3757777864

# Example 2: Mixture of normal distributions
params_mixture <- list(
  mu = c(-1, 1),
  sigma = c(0.5, 0.5),
  lambda = c(0.4, 0.6)
)
logmean_sim_fun(logmean_param = params_mixture, notu = 100)
#>   [1]  1.56591704 -0.98225754  0.73269740  0.87169065  0.63369889  1.48679932
#>   [7]  0.54666608 -1.77808273  1.05408075  1.79153571 -0.64395258  0.85937891
#>  [13]  0.53655529  0.79936222  1.38466041  2.06674154  1.53836512 -1.18286764
#>  [19] -1.61442787  1.19130144  1.25086879  1.50996717  1.22997832 -1.65620021
#>  [25]  0.69130996  0.98802219 -0.77512694 -0.90593007  0.96806365  1.13691199
#>  [31] -1.52727607 -0.33999805 -0.18342352 -1.06050366  1.58473637  0.85076233
#>  [37]  0.59930665  1.48410550 -0.90205090 -0.82344824 -1.36521417  0.26615325
#>  [43] -2.01957072 -0.67336171  0.78729513 -0.21265087  0.71397519 -0.63299620
#>  [49] -1.21577930  1.22395373  1.73468251 -1.10072651  0.99800221  1.16716005
#>  [55]  0.23937903  1.27286472 -0.90483671 -1.32780111 -0.76567792 -0.62825527
#>  [61] -0.82263237  0.35566213 -0.53217074 -1.32932260  0.02934445 -1.75256592
#>  [67]  1.31431309  0.91541259 -0.59540105  0.58430402  0.83701292  0.76869045
#>  [73]  0.14208251 -1.09868024 -1.98729776  1.12268931  1.69772954  0.32285210
#>  [79]  0.86736651  0.16938958  0.92443795  1.47071591 -1.04255570 -1.18029205
#>  [85]  0.96980481  1.24551520  1.04111993 -1.11237839 -1.82478100  0.70843251
#>  [91]  1.99898626 -1.34714664  0.96844293  1.80302076 -1.37876407  0.17269305
#>  [97] -0.94859062  0.82024234  1.40800594  0.96951857