Simulate Log Means for OTUs

This function generates log means for a specified number of OTUs (Operational Taxonomic Units) based on the provided parameters. If a single mean is specified, the log means are drawn from a normal distribution. If multiple means and corresponding weights are specified, the log means are drawn from a mixture of normal distributions.

Usage

logmean_sim_fun(logmean_param, notu)

Arguments

logmean_param

A list containing the parameters for the distribution:

mu: A single value or a vector of mean(s) for the normal or mixture distribution.

sigma: The standard deviation(s) for the normal or mixture distribution.

lambda: (Optional) A vector of weights for the components of the mixture distribution. Required only if mu has more than one value.

notu

An integer specifying the number of OTUs to simulate.

Value

A numeric vector of simulated log means for the specified number of OTUs.

Examples

# Example 1: Single normal distribution
params_single <- list(mu = 0, sigma = 1)
logmean_sim_fun(logmean_param = params_single, notu = 100)
#>   [1]  0.588804559  0.018840329 -0.128790271  0.752496971 -0.399255906
#>   [6]  1.081400706  0.210597235  0.561122468 -3.235268480 -0.408438489
#>  [11]  0.086857457 -1.233233022 -0.795154691  0.234989829  0.975505185
#>  [16] -0.530451833  1.695697500  0.876304868  0.109421045 -0.635344486
#>  [21] -0.118364286 -0.154237577  2.116618561  1.100128458 -0.819765570
#>  [26] -0.863878467 -0.607813648  0.352841463  1.673934219 -0.095797827
#>  [31] -0.008518563 -0.853967162 -1.385920191 -0.234312127  0.050198725
#>  [36] -0.011763044 -0.287874087  0.004836868 -0.041233812 -0.864532804
#>  [41] -0.461601559  0.643573427  0.083646677 -0.033285172  1.595839010
#>  [46]  1.174637946 -0.558311630  1.201119692  0.709478256 -1.391145646
#>  [51]  1.485382867  0.100518870  0.177750365  1.206317686 -0.577546719
#>  [56]  2.570748407  0.645752073  0.915836534  0.265288537 -0.816880567
#>  [61]  0.664690047 -0.613967752 -0.623195126 -0.892805087  0.343857366
#>  [66]  0.512134788  0.108372727 -0.048747047 -0.091832450 -1.710781319
#>  [71]  1.459615281 -1.546137673  1.074037870 -0.703739042  0.894174528
#>  [76] -1.791556874 -0.298701440 -0.485264045  0.546933877  0.157692671
#>  [81]  0.201747809  1.189789713  1.476213261 -0.072073716 -0.587393779
#>  [86]  0.048525665  1.463390773 -0.351330927 -1.857615398  0.139194309
#>  [91]  0.411052658  0.972752901 -1.830893591 -1.147215164 -0.331935125
#>  [96] -0.400009625 -1.454717135 -0.951836000  0.880163256 -0.273236970

# Example 2: Mixture of normal distributions
params_mixture <- list(
  mu = c(-1, 1),
  sigma = c(0.5, 0.5),
  lambda = c(0.4, 0.6)
)
logmean_sim_fun(logmean_param = params_mixture, notu = 100)
#>   [1]  1.07744271 -1.11013574  0.29199882  0.23752221  0.65965639  1.55845995
#>   [7] -0.22890341 -0.58526000  1.00212000  0.78297234  0.22541964  0.81061556
#>  [13]  1.28211600  1.24283393  1.35963503  0.72253582  1.19586581  0.61232017
#>  [19]  1.75995383  0.52364736  1.14851342  1.75705388  1.35994604  1.14359244
#>  [25]  1.39829637 -1.26083780 -1.16414629  0.62887193  0.91880367 -0.46752819
#>  [31]  0.87978830  0.94707690 -1.05944634 -0.73739947 -0.68956245 -1.26346583
#>  [37]  0.66197647 -0.99468868  0.34960863  1.47404840  0.98774757  1.59103912
#>  [43] -0.52717179 -1.02793827 -0.69664963  0.78745479  1.13610278 -0.60488949
#>  [49]  0.96317087  1.24321158  0.90640924  1.34453481  1.87729199 -0.32897706
#>  [55]  1.56078233 -1.25835323 -1.66873834  0.87769701  1.07946100  1.05645775
#>  [61]  0.09311995 -1.20568103  0.95103307 -0.28806749 -0.23346582 -1.28153537
#>  [67]  0.06234719  1.81378310  0.09323532 -1.34175236  0.52057204  0.53741247
#>  [73] -1.51307439  1.21639119 -1.75864429  1.52531795 -0.99508794 -0.57365879
#>  [79] -1.34172861  0.63710936 -0.48854276  0.36986132 -0.63184816 -1.13680878
#>  [85] -1.64210922 -0.89917840 -1.17117605 -0.45470720 -1.80448968  1.11290635
#>  [91]  1.34502294  2.07389505  1.79880209 -1.52891432 -1.10106423 -1.69833116
#>  [97]  0.86992949 -1.21649609  0.97279906  1.56975558