####################################################################### inverse_fun <- function(target,lmb,abs_lfc, model,xmin, xmax) power_fun_ss( deseq_est_list, true_logfoldchange, true_logmean, sample_vec, alpha_level = 0.1, notu ) if (length(target) != length(lmb) || length(target) != length(abs_lfc)) stop("Lengths of 'target', 'lmb', and 'abs_lfc' must be the same.") Initialize an empty list to store the roots roots <- list() roots <- list() roots <- list() Loop through each element and solve for the root for (i in 1:length(target)) roots[i] <- uniroot(function(s) predict(model, type = "response", newdata = data.frame(sample_size = s, lmean_abund = lmbi, abs_lfc = abs_lfci) ) - targeti , interval = c(xmin, xmax), extendInt = "yes")$root unlist(roots) uniroot_lmb = function(target_power,sample_size,abs_lfc,model,xmin,xmax) root <- uniroot(function(lmb) predict(model, type = "response", newdata = data.frame(sample_size = sample_size, lmean_abund = lmb, abs_lfc = abs_lfc) ) - target_power , interval = c(xmin, xmax), extendInt = "yes")$root root inverse_lfc <- function(target,lmb,samp_size, model,xmin, xmax) if (length(target) != length(lmb) || length(target) != length(samp_size)) stop("Lengths of 'target', 'lmb', and 'abs_lfc' must be the same.") for (i in 1:length(target)) roots[i] <- uniroot(function(s) predict(model, type = "response", newdata = data.frame(sample_size = samp_sizei, lmean_abund = lmbi, abs_lfc = s) ) - targeti , interval = c(xmin, xmax), extendInt = "yes")$root unlist(roots) inverse_lmb <- function(target,abs_lfc,samp_size, model,xmin, xmax) if (length(target) != length(abs_lfc) || length(target) != length(samp_size)) stop("Lengths of 'target', 'lmb', and 'abs_lfc' must be the same.") Function to compute the missing input